Remove duplicate conftest files
